unified and simplified test case so project.json is not affected
