+++ src/generator/AutoRest.CSharp.Unit.Tests/SerializationTests.cs
@@ -1,0 +1,19 @@
+﻿using System.IO;
+using AutoRest.Swagger;
+using AutoRest.Swagger.Model;
+using Xunit;
+
+namespace AutoRest.CSharp.Unit.Tests
+{
+    public class SerializationTests
+    {
+        [Fact]
+        public void ParseSecurityDefinitionType()
+        {
+            var swaggerContent = File.ReadAllText(@"Resource\SerializationTests\SerializationTests.yaml");
+            var definition = SwaggerParser.Parse(swaggerContent);
+            Assert.Equal(SecuritySchemeType.OAuth2, definition.SecurityDefinitions["petstore_auth"].SecuritySchemeType);
+            Assert.Equal(SecuritySchemeType.ApiKey, definition.SecurityDefinitions["api_key"].SecuritySchemeType);
+        }
+    }
+}
